RAG/TextSplitter.py
from langchain.text_splitter import RecursiveCharacterTextSplitter

class TextSplitter:
    def __init__(self, chunk_size=1000, chunk_overlap=0):
        self.text_splitter = RecursiveCharacterTextSplitter(
            chunk_size=chunk_size,
            chunk_overlap=chunk_overlap,
            length_function=len
        )

    def get_text_chunks(self, text):
        """Splits text into manageable chunks."""
        return self.text_splitter.split_text(text)