requirements.txt
pyaudio
google-cloud-speech
google-auth
google-cloud-texttospeech
playsound
groq
python-dotenv
SpeechRecognition
fuzzywuzzy
PyQt5
python-Levenshtein
PyMuPDF
openai
langchain-text-splitters
pydantic